IN226063502/RAG_FINAL_PROJECT/requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
langgraph
chromadb
pypdf
python-dotenv
pydantic
tiktoken
